Fixed storybook orcs design theme
