+++ package.json
@@ -77,23 +77,25 @@
     "@mdx-js/react": "^2.3.0",
     "@percy/cli": "^1.26.3",
     "@percy/storybook": "^4.3.6",
-    "@storybook/addon-a11y": "^7.2.1",
-    "@storybook/addon-actions": "^7.2.1",
-    "@storybook/addon-docs": "^7.2.1",
+    "@storybook/addon-a11y": "^7.3.2",
+    "@storybook/addon-actions": "^7.3.2",
+    "@storybook/addon-docs": "^7.3.2",
     "@storybook/addon-knobs": "^7.0.2",
-    "@storybook/addon-links": "^7.2.1",
-    "@storybook/addon-mdx-gfm": "^7.2.1",
-    "@storybook/addon-storyshots": "^7.2.1",
-    "@storybook/addon-toolbars": "^7.2.1",
-    "@storybook/addon-viewport": "^7.2.1",
-    "@storybook/addons": "^7.2.1",
-    "@storybook/api": "^7.2.1",
-    "@storybook/blocks": "^7.2.1",
-    "@storybook/components": "^7.2.1",
+    "@storybook/addon-links": "^7.3.2",
+    "@storybook/addon-mdx-gfm": "^7.3.2",
+    "@storybook/addon-storyshots": "^7.3.2",
+    "@storybook/addon-toolbars": "^7.3.2",
+    "@storybook/addon-viewport": "^7.3.2",
+    "@storybook/addons": "^7.3.2",
+    "@storybook/api": "^7.3.2",
+    "@storybook/blocks": "^7.3.2",
+    "@storybook/components": "^7.3.2",
+    "@storybook/core-events": "^7.3.2",
+    "@storybook/manager-api": "^7.3.2",
     "@storybook/mdx1-csf": "^1.0.0",
-    "@storybook/preset-create-react-app": "^7.2.1",
-    "@storybook/react": "^7.2.1",
-    "@storybook/react-webpack5": "^7.2.1",
+    "@storybook/preset-create-react-app": "^7.3.2",
+    "@storybook/react": "^7.3.2",
+    "@storybook/react-webpack5": "^7.3.2",
     "@storybook/storybook-deployer": "^2.8.16",
     "@testing-library/jest-dom": "^5.11.6",
     "@testing-library/react": "^11.2.2",
@@ -101,7 +103,7 @@
     "@types/styled-components": "^5.1.7",
     "audit-ci": "^6.1.2",
     "babel-core": "^7.0.0-bridge.0",
-    "babel-eslint": "^8.2.5",
+    "babel-eslint": "^10.0.0",
     "babel-loader": "^8.1.0",
     "babel-plugin-react-docgen": "^4.1.0",
     "babel-plugin-styled-components": "^1.10.7",
@@ -120,11 +122,11 @@
     "identity-obj-proxy": "^3.0.0",
     "jest": "^26.6.3",
     "npm-run-all": "^4.1.5",
-    "playroom": "^0.27.7",
+    "playroom": "^0.32.1",
     "prettier": "^2.2.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "storybook": "^7.2.1",
+    "storybook": "^7.3.2",
     "style-loader": "^2.0.0",
     "styled-components": "^5.2.1",
     "webpack": "^5.88.2",
